test: add cases for payment-proof
