+++ src/app/modules/payments/payments.errors.ts
@@ -29,15 +29,3 @@ export class PaymentAccountInformationError extends ApplicationError {
     super(message)
   }
 }
-
-export class PaymentProofUploadS3Error extends ApplicationError {
-  constructor(message = "Can't upload payment proof to S3") {
-    super(message)
-  }
-}
-
-export class PaymentProofPresignS3Error extends ApplicationError {
-  constructor(message = "Can't generate payment proof presign url from S3") {
-    super(message)
-  }
-}
